move test and simplify top comment
